Last week patch: Do or DLC edition (#9)
* Update with the new additions to the game

* Multiple changes

Alright, so... This will be a long one.
Initially, I was going to just implement a map/character select and call it a day this time. But...
First, I decided to fix up the actual fighting and animations some, changing animation speed and animation transitions. A combatant doesn't just move slowly to do something that in reality wouldn't hurt any more than a gentle shove, they punch, hard and fast. And they wouldn't take nearly as long to defend themselves. You two fellow devs are welcome to change it if you feel it too fast, but I feel 1 speed is much too slow to simulate a fight.
Next, I made defending actually do something. To be specific, it lasts less time than being staggered without defending and reduces damage taken by 10 times. Also, now being hit at any time can stagger you, though it still seems just a little delayed.
Then I actually went on to do my initial task, if only for player VS player for now. Options are Ryu, Ken and "Char3", which is currently nothing and would likely break something if selected currently, but we'll implement something in what time we have left.
Map options are numbered without names in the menu right now, because I can't come up with any good names. If we don't come up with anything later, I guess we COULD just name them Bridge, Forest and City, respectively...
However, I ran into an issue in doing so. To do this, I have 3 different scenes, with each end having every character in the game on the exact same spot, but disabled. When characters are selected on the menu, it saves the indexes of the selections in player prefs, then loads the selected scene, activating the selected characters in the scene through minor additions to the HudController script. To get the characters into each of the scenes, I first made the two lists in the bridge map through copying the two usual fighters, then copied and pasted that into every scene, but it seems because of that, battles only progress properly if either Ryu is defeated on the Player 1 side or Ken is defeated on the Player 2 side. Makes me feel like something is hard-coded, but whatever the case really is, I can't seem to find the issue. I'd like the person originally working on the winning code to tell me what the problem is, please, it'd be very appreciated.
I also modified Camera Follow to work with this change, leaving its old version in Player vs AI scenes as Camera Follow Old.